refactor(decimal): move from experimental to components
BREAKING CHANGE: Decimal has moved to src/components/decimal
