+++ src/components/decimal/decimal.spec.js
@@ -3,11 +3,11 @@ import ReactDOM from "react-dom";
 import { mount as enzymeMount } from "enzyme";
 import I18n from "i18n-js";
 
-import { testStyledSystemMargin } from "../../../__spec_helper__/test-utils";
+import { testStyledSystemMargin } from "../../__spec_helper__/test-utils";
 import Decimal from "./decimal.component";
-import Textbox from "../../../components/textbox/textbox.component";
-import Label from "../label";
-import FormFieldStyle from "../form-field/form-field.style";
+import Textbox from "../textbox/textbox.component";
+import Label from "../../__experimental__/components/label";
+import FormFieldStyle from "../../__experimental__/components/form-field/form-field.style";
 
 // These have been written in a way that we can change our testing library or component implementation with relative
 // ease without having to touch the tests.
