+++ src/components/decimal/decimal.stories.js
@@ -3,12 +3,12 @@ import { number, select, boolean, text } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import I18n from "i18n-js";
 import Decimal from "./decimal.component";
-import { getCommonTextboxProps } from "../../../components/textbox/textbox.stories";
-import OptionsHelper from "../../../utils/helpers/options-helper";
-import guid from "../../../utils/helpers/guid";
+import { getCommonTextboxProps } from "../textbox/textbox.stories";
+import OptionsHelper from "../../utils/helpers/options-helper/options-helper";
+import guid from "../../utils/helpers/guid/guid";
 
 export default {
-  title: "Experimental/Decimal Input/Test",
+  title: "Design System/Decimal Input/Test",
   component: Decimal,
   parameters: {
     info: {
