+++ src/components/decimal/decimal.stories.mdx
@@ -1,14 +1,14 @@
 import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import Decimal from ".";
-import { OriginalTextbox } from "../../../components/textbox";
-import Button from "../../../components/button";
-import Box from "../../../components/box";
+import { OriginalTextbox } from "../textbox";
+import Button from "../button";
+import Box from "../box";
 
 <Meta
-  title="Experimental/Decimal Input"
+  title="Design System/Decimal Input"
   parameters={{ info: { disable: true } }}
 />
 
@@ -26,7 +26,7 @@ Captures a number with a decimal point, or a currency value.
 ## Quick Start
 
 ```javascript
-import Decimal from "carbon-react/lib/__experimental__/components/decimal";
+import Decimal from "carbon-react/lib/components/decimal";
 ```
 
 ## Designer Notes
