+++ src/components/inline-inputs/inline-inputs.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Props, Preview, Story } from '@storybook/addon-docs/blocks';
 import InlineInputs from '.';
 import Textbox from '../textbox';
-import Decimal from '../../__experimental__/components/decimal';
+import Decimal from '../decimal';
 import { Select, Option } from '../select';
 import { useState } from 'react';
 import { text, select } from '@storybook/addon-knobs';
